=doc:3472 Updated Stash docs to reflect new mailbox requirements
